factoring out geodata section into geo-picker.html.
step 1:
* new page for picking a geodata record to be used as control of geoid control in main form

Basically working version using the new page as control and getting back the geoid to be mounted to the bound data node. This version uses a repeat instead of a select or datalist for more
flexibility over the content (more than one item can be listed for selection).

Made an icon for the app from original logo.